# internal/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package api

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "BearerAuth.Scopes"
)

// Defines values for CreateProjectRequestStatus.
const (
	CreateProjectRequestStatusActive   CreateProjectRequestStatus = "active"
	CreateProjectRequestStatusArchived CreateProjectRequestStatus = "archived"
	CreateProjectRequestStatusInactive CreateProjectRequestStatus = "inactive"
)

// Defines values for ProjectStatus.
const (
	ProjectStatusActive   ProjectStatus = "active"
	ProjectStatusArchived ProjectStatus = "archived"
	ProjectStatusInactive ProjectStatus = "inactive"
)

// Defines values for UpdateProjectRequestStatus.
const (
	Active   UpdateProjectRequestStatus = "active"
	Archived UpdateProjectRequestStatus = "archived"
	Inactive UpdateProjectRequestStatus = "inactive"
)

// Account defines model for Account.
type Account struct {
	CreatedAt time.Time           `json:"created_at"`
	Email     openapi_types.Email `json:"email"`
	Id        openapi_types.UUID  `json:"id"`
	Name      string              `json:"name"`
	UpdatedAt time.Time           `json:"updated_at"`
}

// AuthResponse defines model for AuthResponse.
type AuthResponse struct {
	AccessToken string  `json:"access_token"`
	Account     Account `json:"account"`

	// ExpiresIn Access token expiration time in seconds
	ExpiresIn    int    `json:"expires_in"`
	RefreshToken string `json:"refresh_token"`
	TokenType    string `json:"token_type"`
}

// CreateProjectRequest defines model for CreateProjectRequest.
type CreateProjectRequest struct {
	Description *string                     `json:"description,omitempty"`
	Name        string                      `json:"name"`
	Status      *CreateProjectRequestStatus `json:"status,omitempty"`
}

// CreateProjectRequestStatus defines model for CreateProjectRequest.Status.
type CreateProjectRequestStatus string

// Error defines model for Error.
type Error struct {
	Error string `json:"error"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Email    openapi_types.Email `json:"email"`
	Password string              `json:"password"`
}

// LogoutRequest defines model for LogoutRequest.
type LogoutRequest struct {
	RefreshToken string `json:"refresh_token"`
}

// Project defines model for Project.
type Project struct {
	AccountId   openapi_types.UUID `json:"account_id"`
	CreatedAt   time.Time          `json:"created_at"`
	Description *string            `json:"description,omitempty"`
	Id          openapi_types.UUID `json:"id"`
	Name        string             `json:"name"`
	Status      ProjectStatus      `json:"status"`
	UpdatedAt   time.Time          `json:"updated_at"`
}

// ProjectStatus defines model for Project.Status.
type ProjectStatus string

// RefreshTokenRequest defines model for RefreshTokenRequest.
type RefreshTokenRequest struct {
	RefreshToken string `json:"refresh_token"`
}

// SignUpRequest defines model for SignUpRequest.
type SignUpRequest struct {
	Email    openapi_types.Email `json:"email"`
	Name     string              `json:"name"`
	Password string              `json:"password"`
}

// UpdateAccountRequest defines model for UpdateAccountRequest.
type UpdateAccountRequest struct {
	Email *openapi_types.Email `json:"email,omitempty"`
	Name  *string              `json:"name,omitempty"`
}

// UpdateProjectRequest defines model for UpdateProjectRequest.
type UpdateProjectRequest struct {
	Description *string                     `json:"description,omitempty"`
	Name        *string                     `json:"name,omitempty"`
	Status      *UpdateProjectRequestStatus `json:"status,omitempty"`
}

// UpdateProjectRequestStatus defines model for UpdateProjectRequest.Status.
type UpdateProjectRequestStatus string

// AccountID defines model for AccountID.
type AccountID = openapi_types.UUID

// ProjectID defines model for ProjectID.
type ProjectID = openapi_types.UUID

// BadRequest defines model for BadRequest.
type BadRequest = Error

// Conflict defines model for Conflict.
type Conflict = Error

// InternalServerError defines model for InternalServerError.
type InternalServerError = Error

// NotFound defines model for NotFound.
type NotFound = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// UpdateAccountJSONRequestBody defines body for UpdateAccount for application/json ContentType.
type UpdateAccountJSONRequestBody = UpdateAccountRequest

// CreateProjectJSONRequestBody defines body for CreateProject for application/json ContentType.
type CreateProjectJSONRequestBody = CreateProjectRequest

// UpdateProjectJSONRequestBody defines body for UpdateProject for application/json ContentType.
type UpdateProjectJSONRequestBody = UpdateProjectRequest

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = LoginRequest

// LogoutJSONRequestBody defines body for Logout for application/json ContentType.
type LogoutJSONRequestBody = LogoutRequest

// RefreshTokenJSONRequestBody defines body for RefreshToken for application/json ContentType.
type RefreshTokenJSONRequestBody = RefreshTokenRequest

// SignUpJSONRequestBody defines body for SignUp for application/json ContentType.
type SignUpJSONRequestBody = SignUpRequest